Popup extraction extra checks
New code seeks footnotes (https://github.com/johnfactotum/foliate/issues/127) in other chapters and returns them as a text for a popup, if they are found. If not, treats them like 'link-internal' and goes to link.

Some epubs (defective or not epub3) can go to an earlier chapter/link, than expected, but that happens even without this popup extraction code.