Generate MD5 hash for FB2 files
Also fixes non-UTF-8 encoded compressed FB2

Note that this will generate the same hash whether the FB2 file is
compressed or not, which is a good thing, I guess? After all, it's still
the very same book if you compress/decompress it. The only drawback is
that you can't just use `md5sum` on the zipped file and match it with
the generated identifier.