Use Tracker, if available, to get file URI
Since Tracker supports extracting identifiers from EPUB books, we can
use the data stored in `~/.local/share/com.github.johnfactotum.Foliate`
as the main source of data to display a book gallery, and then use
Tracker to get the file URIs.

If Tracker is not available or does not have info on a file, then we
will simply fallback on the manual URI store.

See #119