Set icon name on GtkWindow
From the docs:

"This icon is used when the window is minimized (also known as
iconified). Some window managers or desktop environments may also place
it in the window frame, or display it in other contexts. On others,
the icon is not used at all, so your mileage may vary."

Should fix #456