Make debian package compatible with Ubuntu 18.04
For Ubuntu 18.04 the maximum supported version of `debhelper-compat` is 11, while for Ubuntu 20.04 the version it 12. As a result, the package failed to build in Ubuntu 18.04 or 20.04 (due to missing build-dependency). [This article](https://www.man7.org/linux/man-pages/man7/debhelper.7.html) says that `debhelper-compat (=11)` is discouraged, and 12 won't be supported by Ubuntu 18.04, so I used 10 instead.

I have checked that this package builds in both 18.04 and 20.04