Use format native instead of quilt
`quilt` may be used to apply debian specific patches, but it would require creating a upstream tarball during packaging. This commit would save hassle for the upstream developers, when they would distribute their own debian packages.

If the debian specific patches are backported to upstream before submitting to debian, `quilt` won't be required at all. Also, the debian maintainer can use the `quilt` version, but it is really not required in upstream.

Also, if the debian maintainer(s) only change the `foliate/debian` directory, they won't have to apply that as a patch, they can directly edit it. Patches are only to be applied to the rest of the source, if required. In fact, applying patches to debian directory [does not work](https://stackoverflow.com/questions/29634868/adding-a-file-in-a-quilt-dquilt-patch-patch-applies-correctly-by-hand-but-brea/29635229#29635229).