Make debian package compatible with Ubuntu 18.04 and Ubuntu 20.04 (#580)
* Make debian package compatible with Ubuntu 18.04

For Ubuntu 18.04 the maximum supported version of `debhelper-compat` is 11, while for Ubuntu 20.04 the version it 12. As a result, the package failed to build in Ubuntu 18.04 or 20.04 (due to missing build-dependency). [This article](https://www.man7.org/linux/man-pages/man7/debhelper.7.html) says that `debhelper-compat (=11)` is discouraged, and 12 won't be supported by Ubuntu 18.04, so I used 10 instead.

I have checked that this package builds in both 18.04 and 20.04

* Use format native instead of quilt

`quilt` may be used to apply debian specific patches, but it would require creating a upstream tarball during packaging. This commit would save hassle for the upstream developers, when they would distribute their own debian packages.

If the debian specific patches are backported to upstream before submitting to debian, `quilt` won't be required at all. Also, the debian maintainer can use the `quilt` version, but it is really not required in upstream.

Also, if the debian maintainer(s) only change the `foliate/debian` directory, they won't have to apply that as a patch, they can directly edit it. Patches are only to be applied to the rest of the source, if required. In fact, applying patches to debian directory [does not work](https://stackoverflow.com/questions/29634868/adding-a-file-in-a-quilt-dquilt-patch-patch-applies-correctly-by-hand-but-brea/29635229#29635229).