Fix catalogs being emptied after adding/editing catalogs
The bug is caused by the Catalog object being stringified to `{}`.

This must've been due to a recent change in GJS? Because this was
working before. I guess simply `JSON.stringify`ing a GObject object
was sort of a bad idea anyway. Not sure what I was thinking but
anyway, should be fixed now...