Make "more" button on annotation popover show selection menu
The problem was that, if you want to execute actions from the popover,
you need to access the range object again, but you can't because you've
left the webview context.

The solution is a bit convoluted, but I want to avoid recreating the
range from CFI, or storing the range somewhere.

Basically, when showing an annotation or selection popover, the webview
side waits for the value returned by the GJS side. This allows you to
continue execution in the original context.