OPDS: fix vfunc unroot
Also don't set header func. Apparently that crashes GJS. I'm not sure
why the func is being called. It says "Attempting to run a JS callback"
only when closing the last window. It works normally if there are other
windows, or when quitting the app (Ctrl+Q). Anyway, printing the warning
is better than crashing.