OPDS: add CSP
The only thing that this fixes is XSS with `javascript:` links.

Note that `script-src 'self'` does *not* prevent scripts from loading
because origin checks are disabled with `disable-web-security`. This is
okay, though, because there's no way to request scripts; the iframe that
shows the description is sandboxed and all other fields are treated as
plain text.