library: improve fraction calculation
With a zero-based index, this assumes if you are on a
page, that you read it partially exactly to the amount
of pages you read before relative to the amount of pages
before and after. That results in the first page being
considered unread and the last page being fully read of
you were on those pages.

While this is not accurate, there is no way to accurately
determine the current sub-page percentage a person read.
However, at least, this results in the percentage going
from 0 % to 100 % for any given book. Thus, opening a
book for the first time and not starting to read it marks
it as 0 %, which is the desired outcome of this change.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>